requirements/base.txt
# =============================================================================
# JudgeLoom — Base Requirements
# =============================================================================
# Core dependencies shared across all environments.
# =============================================================================

# ─── Framework ──────────────────────────────────────────────────────────────
Django>=5.1,<5.3
django-environ>=0.11.2

# ─── API ────────────────────────────────────────────────────────────────────
django-ninja>=1.3,<2.0
pydantic>=2.9,<3.0
email-validator>=2.2,<3.0

# ─── Database ──────────────────────────────────────────────────────────────
psycopg[binary]>=3.2,<4.0

# ─── Async & Task Queue ───────────────────────────────────────────────────
celery[redis]>=5.4,<6.0
django-celery-beat>=2.7,<3.0

# ─── Real-time ─────────────────────────────────────────────────────────────
channels>=4.1,<5.0
channels-redis>=4.2,<5.0

# ─── Caching ──────────────────────────────────────────────────────────────
django-redis>=5.4,<6.0

# ─── Auth & Security ──────────────────────────────────────────────────────
PyJWT>=2.9,<3.0
django-otp>=1.5,<2.0
qrcode>=7.4,<8.0

# ─── Storage & Media ──────────────────────────────────────────────────────
Pillow>=10.4,<11.0

# ─── Text & Rendering ─────────────────────────────────────────────────────
Markdown>=3.7,<4.0
bleach>=6.1,<7.0
pygments>=2.18,<3.0
lxml>=5.3,<6.0

# ─── Utilities ─────────────────────────────────────────────────────────────
django-extensions>=3.2,<4.0
python-dateutil>=2.9,<3.0
